refactor: put all the User controllers in one file
