--- backend/src/controllers/patchUser.ts
@@ -1,19 +1,0 @@
-import User from '../models/User';
-import { Request, Response } from 'express';
-
-export const patchUser = async (req: Request, res: Response) => {
-  try {
-    const { id } = req.params;
-    const updates = req.body;
-
-    const updatedUser = await User.findByIdAndUpdate(id, updates, { new: true });
-
-    if (!updatedUser) {
-      return res.status(404).json({ message: 'User not found' });
-    }
-
-    res.json({ message: 'User updated', user: updatedUser });
-  } catch (error) {
-    res.status(500).json({ message: 'Error to update user', error });
-  }
-};
